🔨 refactor: 보관함 편지 구조 리팩토링
