+++ src/pages/LetterStoragePage/ReplyStorage/ReplyLetterModal.tsx
@@ -1,11 +1,8 @@
 import { Fragment } from 'react';
-import { toast } from 'react-toastify';
 import { css } from '@emotion/react';
 import Modal from '@/components/Modal';
 import useBoolean from '@/hooks/useBoolean';
 import { Reply } from '@/types/letter';
-import Header from '@/components/Header';
-import { CaretLeft, TrashCan, Copy } from '@/assets/icons';
 import LetterCard from '@/components/LetterCard';
 import LetterHeader from '@/components/LetterHeader';
 import TagList from '@/components/TagList';
@@ -14,68 +11,27 @@ import COLORS from '@/constants/colors';
 import { formatDate } from '@/utils/dateUtils';
 import PolaroidModal from '@/components/PolaroidModal';
 import Button from '@/components/Button';
-import IconButton from '@/components/IconButton';
+import LetterModalHeader from '../components/LetterModalHeader';
 import { getTagList } from '../utils/tagUtills';
-import DeleteBottomSheet from './DeleteBottomSheet';
 
 interface ReplyLetterModalProps extends ReturnType<typeof useBoolean> {
   letter: Reply;
 }
 
 const ReplyLetterModal = ({ value, off, letter }: ReplyLetterModalProps) => {
-  const deleteBottomSheetProps = useBoolean(false);
-
-  const handleContentCopy = (content: string) => {
-    navigator.clipboard
-      .writeText(content)
-      .then(() => {
-        toast.success('편지가 복사되었어요.', {
-          autoClose: 1500,
-          position: 'bottom-center',
-        });
-      })
-      .catch((error) => {
-        console.error(error);
-        toast.error('편지 복사를 실패했어요', {
-          autoClose: 1500,
-          position: 'bottom-center',
-        });
-      });
-  };
-
   return (
     <Modal isOpen={value} close={off}>
-      <Header css={style.header}>
-        <Header.Left>
-          <div css={style.leftHeader}>
-            <CaretLeft
-              css={style.icon}
-              strokeWidth={2.5}
-              color="white"
-              onClick={off}
-            />
-            <span css={style.headerText}>보관한 편지</span>
-          </div>
-        </Header.Left>
-        <Header.Right>
-          <div css={style.rightHeader}>
-            <IconButton
-              onClick={() => {
-                handleContentCopy(
-                  letter.letterType !== 'Onboarding'
-                    ? letter.repliedContent
-                    : letter.content,
-                );
-              }}
-            >
-              <Copy stroke="white" />
-            </IconButton>
-            <IconButton onClick={deleteBottomSheetProps.on}>
-              <TrashCan fill="white" />
-            </IconButton>
-          </div>
-        </Header.Right>
-      </Header>
+      <LetterModalHeader
+        headerText="보관한 편지"
+        copyContent={
+          letter.letterType === 'Onboarding'
+            ? letter.content
+            : letter.repliedContent
+        }
+        letterId={letter.letterId}
+        off={off}
+        type="reply"
+      />
       <div css={style.container}>
         <LetterCard css={style.card(letter.letterType)}>
           {letter.letterType === 'Onboarding' ? (
@@ -132,11 +88,6 @@ const ReplyLetterModal = ({ value, off, letter }: ReplyLetterModalProps) => {
           )}
         </LetterCard>
       </div>
-      <DeleteBottomSheet
-        letterId={letter.letterId}
-        modalOff={off}
-        {...deleteBottomSheetProps}
-      />
     </Modal>
   );
 };
@@ -144,30 +95,6 @@ const ReplyLetterModal = ({ value, off, letter }: ReplyLetterModalProps) => {
 export default ReplyLetterModal;
 
 const style = {
-  header: css`
-    height: 2.5rem;
-    margin-bottom: 0.5rem;
-    padding-top: 1.25rem;
-    padding-bottom: 0.5rem;
-  `,
-  leftHeader: css`
-    display: flex;
-    gap: 0.625rem;
-    align-items: center;
-  `,
-  rightHeader: css`
-    display: flex;
-    gap: 0.625rem;
-  `,
-  icon: css`
-    cursor: pointer;
-  `,
-  headerText: css`
-    color: white;
-    font-weight: 600;
-    font-size: 16px;
-    line-height: 16px;
-  `,
   container: css`
     overflow-y: auto;
     max-height: calc(100svh - 60px);
