+++ src/pages/LetterStoragePage/hooks/useLetterReply.ts
@@ -1,7 +1,7 @@
 import { useSuspenseQuery } from '@tanstack/react-query';
 import letterOptions from '@/api/letter/queryOptions';
 
-const useLetterStorage = (page: number) => {
+const useLetterReply = (page: number) => {
   const { data } = useSuspenseQuery({
     ...letterOptions.storage(page.toString()),
   });
@@ -9,4 +9,4 @@ const useLetterStorage = (page: number) => {
   return data;
 };
 
-export default useLetterStorage;
+export default useLetterReply;
